website/requirements.txt
# Web framework & server
fastapi==0.103.1
uvicorn==0.23.2

# Templating & env
jinja2==3.1.2
python-dotenv==1.0.0

#Auth & security
python-multipart==0.0.6

# Data & ML
numpy==1.24.3
pandas>=1.5.3,<2.0

# Extras & utilities
openpyxl==3.1.5
python-engineio==4.3.4
psycopg2-binary==2.9.10
requests==2.28.2
asyncpg==0.26.0
itsdangerous>=2.1.2
PyYAML


